Refactor resetFilters to check for filterName input
The issue isn't with your Google Sheet link but with a small typo in the JavaScript code. The console is showing an Uncaught SyntaxError, which is a critical error that stops the browser from running any more of your script. This is why the data fetching and card rendering never happen.

The Main Problem: Syntax Error ️
The error is on line 407 of your code. You have a ! character that is not valid in plain JavaScript. This syntax is used in TypeScript but will cause a crash when run directly in a browser.

File: (index)

Line: 407

Error: Uncaught SyntaxError: Unexpected token '!'

How to Fix It
You just need to remove the ! from that line. For extra safety, it's best to also check if the element exists before trying to change its value.

Find this line in your <script> tag:

JavaScript

$('#filterName')!.value='';
Replace it with this corrected code:

JavaScript

const filterNameInput = $('#filterName');
if (filterNameInput) {
  filterNameInput.value = '';
}
This change removes the syntax error and safely resets the input field's value. Once you make this fix, your script should run correctly and the player cards will load.

const filterNameInput = $('#filterName');
if (filterNameInput) {
  filterNameInput.value = '';
}