Update alumni-news.html
trying to find the last version of code that rendered the line scores.  many subsequent changes to other parts of the page have already been changed, so I realize that by going backwards we will lose those changes,but i need to uncover what caused the linescores to disappear forever.